parsing fait

+++ philo/sources/write_state.c
@@ -6,33 +6,33 @@
 /*   By: chhoflac <chhoflac@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/22 15:41:28 by chhoflac          #+#    #+#             */
-/*   Updated: 2024/05/24 11:03:42 by chhoflac         ###   ########.fr       */
+/*   Updated: 2024/05/24 15:22:07 by chhoflac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-void	take_fork(double time_stamp, int philosopher)
+void	take_fork(int time_stamp, int philosopher)
 {
 	printf("%d %d has taken a fork", time_stamp, philosopher);
 }
 
-void	is_eating(double time_stamp, int philosopher)
+void	is_eating(int time_stamp, int philosopher)
 {
 	printf("%d %d is eating", time_stamp, philosopher);
 }
 
-void	is_sleeping(double time_stamp, int philosopher)
+void	is_sleeping(int time_stamp, int philosopher)
 {
 	printf("%d %d is sleeping", time_stamp, philosopher);
 }
 
-void	is_thinking(double time_stamp, int philosopher)
+void	is_thinking(int time_stamp, int philosopher)
 {
 	printf("%d %d is thinking", time_stamp, philosopher);
 }
 
-void	died(double time_stamp, int philosopher)
+void	died(int time_stamp, int philosopher)
 {
 	printf("%d %d died", time_stamp, philosopher);
 }
